tested a bit

+++ local/ft_printf/code/libftprintf.h
@@ -10,7 +10,7 @@ int				arg_to_str(const char *);
 int				arg_to_ptr(unsigned long long ptr);
 int				arg_to_int(int d);
 unsigned int	arg_to_uint(unsigned int n);
-void			int_to_hexadecimal(unsigned long long n, int uppercase);
+void			int_to_hexadecimal(int n, int uppercase);
 int				calculate_int_length(int n);
 int				arg_to_hexa(int n, int uppercase);
